+++ src/utility.h
@@ -9,9 +9,9 @@
 
 std::pair<glm::vec3, glm::vec3> min_max_of(std::span<glm::vec3 const>);
 
-std::pair<glm::vec3, glm::vec3> min_max_of(std::span<double const> x,
-                                           std::span<double const> y,
-                                           std::span<double const> z);
+std::pair<glm::vec3, glm::vec3> min_max_of(std::span<float const> x,
+                                           std::span<float const> y,
+                                           std::span<float const> z);
 
 void update_instances(std::span<glm::mat4 const> instances,
                       noo::DocumentTPtr          doc,
